+++ cms/src/pages/editPost/editPostLoader.ts
@@ -2,21 +2,17 @@ import { ActionFunctionArgs, ParamParseKey, Params } from "react-router";
 import { axiosPrivate } from "../../config/axios.config";
 import { Post } from "../../interfaces/blog";
 
-/* How to type params for route loaders?
- * https://stackoverflow.com/questions/75324193/react-router-6-how-to-strongly-type-the-params-option-in-route-loader
- */
-
 const Paths = {
-  postDetail: "/:username/:postTitle",
+  editPostDetail: "/:username/posts/:postTitle/edit",
 } as const;
 
 interface LoaderArgs extends ActionFunctionArgs {
-  params: Params<ParamParseKey<typeof Paths.postDetail>>;
+  params: Params<ParamParseKey<typeof Paths.editPostDetail>>;
 }
 
-const postLoader = async ({ params }: LoaderArgs): Promise<Post> => {
+const editPostLoader = async ({ params }: LoaderArgs): Promise<Post> => {
   console.group("postLoader running...");
-  console.log("params:", params);
+
   try {
     const url = `users/${params.username}/posts/${params.postTitle}`;
     const response = await axiosPrivate.get(url);
@@ -29,4 +25,4 @@ const postLoader = async ({ params }: LoaderArgs): Promise<Post> => {
   }
 };
 
-export default postLoader;
+export default editPostLoader;